[BORA00015262] wlan gen4m: Printf arg type mismatch
[Detail]
Printf arg type mismatch:
%u <- (prAdapter->rVerInfo.u2FwOwnVersion >> 8)

[Solution]
%u <- (uint16_t)(prAdapter->rVerInfo.u2FwOwnVersion >> 8)

MTK-Commit-Id: e2a5760af3ebec037ac5abd80fb04034c8256f20

CR-Id: BORA00015262
Feature: Others
Reviewed-on: http://gerrit.mediatek.inc:8080/1414449
CheckPatch: Check Patch <srv_checkpatch@mediatek.com>
Build: srv_neptune_adm <srv_neptune_adm@mediatek.com>
Test: srv_mspautosanity@mediatek.com

+++ os/linux/gl_rst.c
@@ -314,13 +314,15 @@ u_int8_t glResetTrigger(struct ADAPTER *prAdapter,
 		       MTK_CHIP_REV, wlanGetEcoVersion(prAdapter));
 		DBGLOG(INIT, ERROR,
 		       "FW Ver DEC[%u.%u] HEX[%x.%x], Driver Ver[%u.%u]\n",
-		       (prAdapter->rVerInfo.u2FwOwnVersion >> 8),
-		       (prAdapter->rVerInfo.u2FwOwnVersion & BITS(0, 7)),
-		       (prAdapter->rVerInfo.u2FwOwnVersion >> 8),
-		       (prAdapter->rVerInfo.u2FwOwnVersion & BITS(0, 7)),
-		       (prAdapter->rVerInfo.u2FwPeerVersion >> 8),
-		       (prAdapter->rVerInfo.u2FwPeerVersion & BITS(0, 7)));
-
+		       (uint16_t)(prAdapter->rVerInfo.u2FwOwnVersion >> 8),
+		       (uint16_t)(prAdapter->rVerInfo.u2FwOwnVersion &
+				  BITS(0, 7)),
+		       (uint16_t)(prAdapter->rVerInfo.u2FwOwnVersion >> 8),
+		       (uint16_t)(prAdapter->rVerInfo.u2FwOwnVersion &
+				  BITS(0, 7)),
+		       (uint16_t)(prAdapter->rVerInfo.u2FwPeerVersion >> 8),
+		       (uint16_t)(prAdapter->rVerInfo.u2FwPeerVersion &
+				  BITS(0, 7)));
 		fgResult = TRUE;
 	} else {
 		DBGLOG(INIT, ERROR,
@@ -328,13 +330,15 @@ u_int8_t glResetTrigger(struct ADAPTER *prAdapter,
 		       pucFile, u4Line,
 		       MTK_CHIP_REV,
 		       wlanGetEcoVersion(prAdapter),
-		       (prAdapter->rVerInfo.u2FwOwnVersion >> 8),
-		       (prAdapter->rVerInfo.u2FwOwnVersion & BITS(0, 7)),
-		       (prAdapter->rVerInfo.u2FwOwnVersion >> 8),
-		       (prAdapter->rVerInfo.u2FwOwnVersion & BITS(0, 7)),
-		       (prAdapter->rVerInfo.u2FwPeerVersion >> 8),
-		       (prAdapter->rVerInfo.u2FwPeerVersion & BITS(0, 7)));
-
+		       (uint16_t)(prAdapter->rVerInfo.u2FwOwnVersion >> 8),
+		       (uint16_t)(prAdapter->rVerInfo.u2FwOwnVersion &
+				  BITS(0, 7)),
+		       (uint16_t)(prAdapter->rVerInfo.u2FwOwnVersion >> 8),
+		       (uint16_t)(prAdapter->rVerInfo.u2FwOwnVersion &
+				  BITS(0, 7)),
+		       (uint16_t)(prAdapter->rVerInfo.u2FwPeerVersion >> 8),
+		       (uint16_t)(prAdapter->rVerInfo.u2FwPeerVersion &
+				  BITS(0, 7)));
 		wifi_rst.rst_trigger_flag = u4RstFlag;
 		schedule_work(&(wifi_rst.rst_trigger_work));
 	}
